This patch fixes a bug reported by Graham Leggett (minfrin@sharp.fm). The bug was that any insert or update would fail if the returned oid was larger than a signed int.  Since OIDs are unsigned int's it was a bug that the code used a java signed int to deal with the values.  The bug would result in the error message: "Unable to fathom update count". While fixing the bug, it became apparent that other code made a similar assumption about OIDs being signed ints.  Therefore some methods that returned or took OIDs are arguements also needed to be changed. Since we are so close to the 7.2 release I have added new methods that return longs and deprecated the old methods returning ints.  Therefore all old code should still work without requiring a code change to cast from long to int.  Also note that the methods below are PostgreSQL specific extensions to the JDBC api are are not part of the spec from Sun, thus it is unlikely that they are used much or at all.
The deprecated methods are:
  ResultSet.getInsertedOID()
  Statement.getInsertedOID()
  Serialize.store()
  Connection.putObject()
and are replaced by:
  ResultSet.getLastOID()
  Statement.getLastOID()
  Serialize.storeObject()
  Connection.storeObject()
All the deprecated methods returned int, while their replacements return long

This patch also fixes two comments in MD5Digest that the author Jeremy Wohl
submitted.

--Barry

+++ src/interfaces/jdbc/org/postgresql/core/QueryExecutor.java
@@ -13,7 +13,7 @@ import org.postgresql.util.PSQLException;
  * <p>The lifetime of a QueryExecutor object is from sending the query
  * until the response has been received from the backend.
  *
- * $Id: QueryExecutor.java,v 1.5 2001/11/19 23:16:45 momjian Exp $
+ * $Id: QueryExecutor.java,v 1.6 2001/11/25 23:26:56 barry Exp $
  */
 
 public class QueryExecutor
@@ -46,7 +46,7 @@ public class QueryExecutor
 	private boolean binaryCursor = false;
 	private String status = null;
 	private int update_count = 1;
-	private int insert_oid = 0;
+	private long insert_oid = 0;
 	private int maxRows;
 
 	/*
@@ -173,7 +173,7 @@ public class QueryExecutor
 			}
 			if (status.startsWith("INSERT"))
 			{
-				insert_oid = Integer.parseInt(status.substring(1 + status.indexOf(' '),
+				insert_oid = Long.parseLong(status.substring(1 + status.indexOf(' '),
 											  status.lastIndexOf(' ')));
 			}
 		}
